Finalize dataset structure and fix minor bugs
